Restructure and redesign the Lua compiler.

+++ luacompiler/src/register_map.rs
@@ -4,8 +4,8 @@ use std::collections::HashMap;
 /// Represents a structure that is used to keep track of the mapping
 /// between Lua variables and register ids.
 pub struct RegisterMap {
-    reg_count: RefCell<usize>,
-    reg_map: RefCell<HashMap<String, usize>>
+    reg_count: RefCell<u8>,
+    reg_map: RefCell<HashMap<String, u8>>
 }
 
 impl RegisterMap {
@@ -19,7 +19,7 @@ impl RegisterMap {
     /// Generates a fresh register and returns it.
     /// This is used in cases like `x = 1 + 2 + 3` to generate intermmediate
     /// registers in which, for instance, the result of 2 + 3 is stored.
-    pub fn new_reg(&self) -> usize {
+    pub fn new_reg(&self) -> u8 {
         let to_return = *self.reg_count.borrow();
         *self.reg_count.borrow_mut() += 1;
         to_return
@@ -28,13 +28,13 @@ impl RegisterMap {
     /// Get the register that corresponds to the given identifier.
     /// If the corresponding register is not found, a new register is created
     /// and returned.
-    pub fn get_reg(&self, name: &str) -> usize {
+    pub fn get_reg(&self, name: &str) -> u8 {
         *self.reg_map.borrow_mut().entry(name.to_string())
             .or_insert_with(|| self.new_reg())
     }
 
     /// Get the total number of registers that were needed.
-    pub fn reg_count(self) -> usize {
+    pub fn reg_count(self) -> u8 {
         *self.reg_count.borrow_mut()
     }
 }
